refactor: almost the entire API was moved from the templ package to the templ_internal package. the runtime_test.go has been moved to templ_internal and it's passing.
note: another thing that i did was to generate the .templ files in this repo with the new generator so that is the reason of this big commit but I only changed the generator package and the runtime.go files and tests.

unnecesary code or bloatware removed: the functions that are unrelated to html, css and js templating were removed.