fix: include all watch patterns in consideration for cmd restart
When -watch-pattern is set and a non-go and non-text (templ text) file changes, processing the change is effectively a no-op because templ only takes action on changes to go files and text files. However, there are scenarious when changes to any watch file should cause the 

This changes causes templ to re-run its "cmd" argument and time files matching its -watch-pattern change.